Extract functions for mapping nodes and edges to separate functions
